+++ utils/hosp-poll.c
@@ -9,6 +9,7 @@
 #include <signal.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <hidapi.h>
 #include <hosp.h>
 #include "util.h"
 
@@ -167,8 +168,14 @@ int main(int argc, char** argv) {
   signal(SIGINT, shandle);
   parse_args(argc, argv);
 
+  if (hid_init() < 0) {
+    fprintf(stderr, "hid_init: %ls\n", hid_error(NULL));
+    return 1;
+  }
+
   if ((hosp = hosp_open()) == NULL) {
     perror("Failed to open ODROID Smart Power connection");
+    hid_exit();
     return errno;
   }
 
@@ -181,5 +188,6 @@ int main(int argc, char** argv) {
     perror("Failed to close ODROID Smart Power connection");
   }
 
+  hid_exit();
   return ret;
 }