+++ utils/hosp-set.c
@@ -8,6 +8,7 @@
 #include <getopt.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <hidapi.h>
 #include <hosp.h>
 #include "util.h"
 
@@ -96,8 +97,14 @@ int main(int argc, char** argv) {
 
   parse_args(argc, argv);
 
+  if (hid_init() < 0) {
+    fprintf(stderr, "hid_init: %ls\n", hid_error(NULL));
+    return 1;
+  }
+
   if ((hosp = hosp_open()) == NULL) {
     perror("Failed to open ODROID Smart Power connection");
+    hid_exit();
     return errno;
   }
 
@@ -108,5 +115,6 @@ int main(int argc, char** argv) {
     perror("Failed to close ODROID Smart Power connection");
   }
 
+  hid_exit();
   return ret;
 }