+++ RELEASES.md
@@ -8,8 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- hosp_enumerate: function to wrap `hid_enumerate`.
+- hosp_open_device: function to open using an existing `hid_device`.
+- hosp_get_device: function to get the underlying `hid_device` handle.
+
 ### Changed
 
+- HIDAPI is now a public dependency (was private).
 - Build:
   - Increase minimum required CMake version to 3.16.
   - Prefer native HIDAPI backends over `libusb` backend.